+++ ecs_crd/cli.py
@@ -67,14 +67,14 @@ class Parameters:
 def main():
     pass
 
-@main.command(help='validate the configuration file')
+@main.command(help='dry run the configuration file')
 @click.option('-e', '--environment', required=True, default='stage', help='Environment to deploy.', show_default=True)
 @click.option('-r', '--region', required=True, default='eu-west-3', help='Amazon Web Service region used to deploy ECS service.', show_default=True)
 @click.option('-f', '--configuration-file', required=False, help='deployment configuration file.')
 @click.option('-d', '--configuration-dir', required=False, help='directory to find the deployment configuration file.')
 @click.option('--verbose', is_flag=True, default=False, help='activate verbose log.')
 @click.option('--log-file', required=False, help='output log file result.')
-def validate(
+def dry_run(
         environment,
         region,
         configuration_file,